+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2023-11-01
+
 ### Added
 
 - Role in player profile
@@ -36,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[unreleased]: https://github.com/riftlink/riftlink-frontend/compare/v1.1.0...HEAD
-[1.0.1]: https://github.com/riftlink/riftlink-frontend/compare/v1.0.0...v1.1.0
+[unreleased]: https://github.com/riftlink/riftlink-frontend/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/riftlink/riftlink-frontend/compare/v1.1.0...v1.2.0
+[1.1.0]: https://github.com/riftlink/riftlink-frontend/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/riftlink/riftlink-frontend/releases/tag/v1.0.0